+++ src/main/java/com/junsu/cyr/controller/chat/ChatRoomUserController.java
@@ -1,7 +1,6 @@
 package com.junsu.cyr.controller.chat;
 
-import com.junsu.cyr.flow.chat.ExitChatRoomFlow;
-import com.junsu.cyr.flow.chat.JoinChatRoomFlow;
+import com.junsu.cyr.service.chat.ChatService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -11,18 +10,17 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/chat/room/user")
 public class ChatRoomUserController {
 
-    private final JoinChatRoomFlow joinChatRoomFlow;
-    private final ExitChatRoomFlow exitChatRoomFlow;
+    private final ChatService chatService;
 
     @PostMapping("/{chatRoomId}")
     public ResponseEntity<String> joinChatRoom(@PathVariable("chatRoomId") Long chatRoomId, @RequestAttribute Integer userId) {
-        joinChatRoomFlow.joinChatRoom(chatRoomId, userId);
+        chatService.firstJoinChatRoom(chatRoomId, userId);
         return ResponseEntity.ok("success to join chat room");
     }
 
     @DeleteMapping("/{chatRoomId}")
     public ResponseEntity<String> exitChatRoom(@PathVariable("chatRoomId") Long chatRoomId, @RequestAttribute Integer userId) {
-        exitChatRoomFlow.exitChatRoom(chatRoomId, userId);
+        chatService.deleteUserFromChatRoom(chatRoomId, userId);
         return ResponseEntity.ok("success to remove user from chat room");
     }
 }
