+++ src/main/java/com/junsu/cyr/domain/chats/ChatRoom.java
@@ -46,9 +46,6 @@ public class ChatRoom extends BaseTime {
 
     public void decreaseMemberCount() {
         memberCount--;
-        if(memberCount < 0) {
-            this.memberCount = 0L;
-        }
     }
 
     public void updateLastMessage(ChatMessage chatMessage) {
