--- src/main/java/com/junsu/cyr/global/annotation/ManagerOnly.java
@@ -1,11 +1,0 @@
-package com.junsu.cyr.global.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.METHOD)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface ManagerOnly {
-}
