+++ src/main/java/com/junsu/cyr/service/chat/ChatMessageService.java
@@ -8,6 +8,7 @@ import com.junsu.cyr.model.chat.ChatMessageConditionRequest;
 import com.junsu.cyr.model.chat.ChatMessageResponse;
 import com.junsu.cyr.repository.ChatMessageRepository;
 import com.junsu.cyr.response.exception.http.BaseException;
+import com.junsu.cyr.response.exception.code.ChatMessageExceptionCode;
 import com.junsu.cyr.response.exception.code.ChatRoomUserExceptionCode;
 import com.junsu.cyr.service.user.UserNicknameSettingService;
 import com.junsu.cyr.service.user.UserService;
@@ -30,6 +31,11 @@ public class ChatMessageService {
     private final ChatRoomUserService chatRoomUserService;
     private final UserNicknameSettingService userNicknameSettingService;
 
+    public ChatMessage getChatMessageByChatRoomId(Long chatMessageId) {
+        return chatMessageRepository.findById(chatMessageId)
+                .orElseThrow(() -> new BaseException(ChatMessageExceptionCode.NOT_FOUND_CHAT_ROOM));
+    }
+
     @Transactional
     public ChatMessage createUserMessage(ChatRoom chatRoom, User user, String content, Type type) {
         ChatMessage chatMessage = ChatMessage.builder()
@@ -39,7 +45,6 @@ public class ChatMessageService {
                 .type(type)
                 .build();
 
-        chatRoom.updateLastMessage(chatMessage);
         return chatMessageRepository.save(chatMessage);
     }
 
@@ -52,7 +57,6 @@ public class ChatMessageService {
                 .type(Type.SYSTEM)
                 .build();
 
-        chatRoom.updateLastMessage(chatMessage);
         return chatMessageRepository.save(chatMessage);
     }
