+++ src/main/java/com/junsu/cyr/service/event/EventFlowService.java
@@ -1,12 +1,9 @@
-package com.junsu.cyr.flow.event;
+package com.junsu.cyr.service.event;
 
 import com.junsu.cyr.domain.events.Event;
 import com.junsu.cyr.domain.users.User;
-import com.junsu.cyr.model.event.EventResponse;
-import com.junsu.cyr.model.event.EventUploadRequest;
 import com.junsu.cyr.response.exception.code.UserExceptionCode;
 import com.junsu.cyr.response.exception.http.BaseException;
-import com.junsu.cyr.service.event.EventService;
 import com.junsu.cyr.service.user.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -14,20 +11,22 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
-public class CreateEventFlow {
+public class EventFlowService {
 
-    private final UserService userService;
     private final EventService eventService;
+    private final EventCommentService eventCommentService;
+    private final UserService userService;
 
     @Transactional
-    public EventResponse createEvent(EventUploadRequest request, Integer userId) {
+    public void deleteEvent(Long eventId, Integer userId) {
         User user = userService.getUserById(userId);
 
         if(!userService.isLeastManager(user)) {
             throw new BaseException(UserExceptionCode.REQUIRES_AT_LEAST_MANAGER);
         }
 
-        Event event = eventService.createEvent(request, user);
-        return new EventResponse(event);
+        Event event = eventService.findEventByEventId(eventId);
+        eventCommentService.deleteAllEventComment(event);
+        eventService.deleteEvent(event);
     }
 }
